+++ src/buttercat/Main.java
@@ -1,21 +1,5 @@
 package buttercat;
 
-import TI.BoeBot;
-import hardwarelayer.sensors.button.Button;
-import hardwarelayer.sensors.infraredreceiver.InfraRedReceiver;
-import hardwarelayer.sensors.linefollower.LineFollower;
-import interfacelayer.Engine;
-import interfacelayer.LineFollowChecker;
-import interfacelayer.NotificationSystem;
-import interfacelayer.gui.PathVisualiser;
-import javafx.application.Application;
-import ooframework.FrameworkProgram;
-import TI.SerialConnection;
-
-import TI.SerialConnection;
-
-import java.util.ArrayList;
-
 
 public class Main {
     public static void main(String[] args) {
