+++ src/interfacelayer/gui/grid/PathVisualiser.java
@@ -1,14 +1,9 @@
-package interfacelayer.gui;
+package interfacelayer.gui.grid;
 
 import javafx.application.Application;
-import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.GridPane;
-import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
 
 public class PathVisualiser extends Application {
@@ -21,7 +16,8 @@ public class PathVisualiser extends Application {
 
     @Override
     public void start(Stage stage) throws Exception {
-        Parent root = FXMLLoader.load(getClass().getResource("pathLayout.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("layout.fxml"));
+
 
         Scene scene = new Scene(root);
         scene.getStylesheets().add("interfacelayer/gui/stylesheet.css");
