+++ src/interfacelayer/gui/stylesheet.css
@@ -19,3 +19,17 @@
     -fx-text-fill: #f0f2f0;
     -fx-background-color: #5dba62;
 }
+
+#helpButton {
+    -fx-text-fill: #f0f2f0;
+    -fx-background-color: #5dba62;
+}
+
+#typeBox{
+    
+    -fx-background-color: #5dba62;
+}
+
+.combo-box:text-input {
+    -fx-text-fill: #f0f2f0; 
+}
